refactor(tests): split asset integration boundaries
