+++ packages/host/assets/tests/integration/download.test.ts
@@ -4,11 +4,11 @@ import { tmpdir } from 'node:os';
 import { join } from 'node:path';
 import type { ManagedAssetArtifact } from '@linkcode/schema';
 import { afterEach, describe, expect, it } from 'vitest';
-import type { LocalServer } from '../../src/__tests__/helpers/local-server';
-import { startLocalServer } from '../../src/__tests__/helpers/local-server';
 import type { DownloadProgress } from '../../src/download';
 import { downloadVerified } from '../../src/download';
 import { DownloadError, IntegrityError } from '../../src/errors';
+import type { LocalServer } from './helpers/local-server';
+import { startLocalServer } from './helpers/local-server';
 
 const payload = randomBytes(256 * 1024);
 const sri = `sha512-${createHash('sha512').update(payload).digest('base64')}`;
