+++ packages/host/assets/tests/integration/install.test.ts
@@ -2,15 +2,15 @@ import { mkdtempSync, readdirSync, readFileSync, statSync } from 'node:fs';
 import { tmpdir } from 'node:os';
 import { dirname, join } from 'node:path';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import type { TgzFixture } from '../../src/__tests__/helpers/fixtures';
-import { makeTgz } from '../../src/__tests__/helpers/fixtures';
-import type { LocalServer } from '../../src/__tests__/helpers/local-server';
-import { startLocalServer } from '../../src/__tests__/helpers/local-server';
 import type { AssetDescriptor } from '../../src/catalog';
 import { IntegrityError } from '../../src/errors';
 import { installAsset, installedComplete, installedPath } from '../../src/install';
 import { assetDir } from '../../src/paths';
 import { currentPlatformKey } from '../../src/platform';
+import type { TgzFixture } from './helpers/fixtures';
+import { makeTgz } from './helpers/fixtures';
+import type { LocalServer } from './helpers/local-server';
+import { startLocalServer } from './helpers/local-server';
 
 const platform = currentPlatformKey();
 if (!platform) throw new Error('tests require a catalog-supported platform');
